refactor files in build_soma
